foam wip

+++ Assets/Water/Resources/WaterHelperLib.cginc
@@ -69,7 +69,7 @@ float2 gaussian_pair(float2 n, float time)
 }
 
 
-float PhilipSpectrum(float amplitude, double2 wavevector, float2 wind, float gravity) 
+float PhilipSpectrum(float amplitude, float2 wavevector, float2 wind, float gravity) 
 {
 	float2 waveDir = normalize(wavevector);
 	float2 windDir = normalize(wind);
